Start with a reader and composite plotter
Implement a scene reader from the USGS EarthExplorer download format.
Works for decompressed bands in a folder or a tar archive with the
bands. Reads the metadata about conversions and coordinates from the
*MTL.txt file and puts everything into an xarray.Dataset for easy access
and plotting. Includes a function to plot a composite from the scene.
Missing tests and proper docstrings.