Add the save_scene function to output back to disk (#19)
This function can be used to same a scene back into the EarthExplorer
format. The output is written directly to a tar file (with or without
compression).  It could be useful when cropping a scene for a class or 
the docs. Add a warning that for general storage after a scene is 
loaded, using `xarray.Dataset.to_netcdf` is better.