Create a datasets module to automate downloading (#24)
The sample data download functions make it easier to fetch the sample
data without having to type the DOI, MD5, and file name all the time.
Add an option to untar the archive after fetching to make it easier to
test IO functions. Use the download HTTPS url directly to avoid issues
with Pooch and figshare DOI versions.