Use np.abs in pansharpen to avoid sign issues
The band average calculation and multiplication with the pan band from
the original method assumes we're operating on the DN values, which are
all positive. But since we do this on reflectance instead, we need to
take absolute values when doing these to avoid issues with signs
flipping in the output.