Fetch data files from GitHub instead of figshare
Avoid hitting their servers all the time from CI. For github, at least
the data has a chance of being close to the CI. Transition to using a
Pooch class and wrap repeated unpacking code into a function. Had to
allow setuptools-scm to set a local version of Pooch to use the "main"
branch correctly.