Add a trigger warning and update the indices tutorial (#53)
Using the Brumadinho data is a bit tricky because a lot of people died
during the event. But the data are perfect for the example we want to
demonstrate. So add a bit more context to acknowledge the tragedy and a
trigger warning. Also update the tutorial with more information.