Add NDVI calculation
Add a function that calculates NDVI for a given scene. It's better than
having to remember the equations all the time. The function can also
take care of adding some metadata and copying the metadata from the
original scene. The function allows customizing the names of the red and
NIR bands.